min max perfomance

+++ mem/filearchive.h
@@ -63,7 +63,7 @@ public:
 	OArchive& save_bin(const void* ptr, size_t size) { pos += size; return *this;}
 	size_t opos() const {return pos;}
 private:
-	size_t pos;
+	uint64_t pos;
 };
 
 }//namespace
